pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-prep"
version = "0.1.0"
description = "Local AI stack: CatBoost + RAG + FastAPI + MLflow + DuckDB"
readme = "README.md"
authors = [{ name = "Robert Leggett", email = "you@example.com" }]
requires-python = ">=3.11"

dependencies = [
    # Core
    "pandas>=2.3.3",
    "numpy>=2.3.4",
    "scikit-learn>=1.7.2",
    "pyyaml>=6.0.3",
    "urllib3<2.5.0",
    "requests<2.32.5",

    # Classic ML
    "catboost>=1.2.8",
    "shap>=0.49.1",

    # Tracking/registry
    "mlflow>=3.5.1",
    "cloudpickle>=3.1.2",

    # API
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.4",

    # Sever
    "mcp>=1.21.0",

    # RAG + embeddings
    "chromadb>=1.3.4",
    "sentence-transformers>=5.1.2",
    "faiss-cpu>=1.12.0",

    # LLMs & agents
    "langchain>=1.0.0",
    "langchain-ollama>=0.3.9",
    "langgraph>=0.6.11",
    "openai>=2.7.1",

    # Lakehouse
    "duckdb>=1.4.1",
    "pyarrow>=21.0.0",

    # Ingestion & tokens
    "pypdf>=6.1.3",
    "tiktoken>=0.12.0",

    # Optional utils
    "python-dotenv>=1.2.1",
    "joblib>=1.5.2",

    # Notebooks & viz (optional)
    "jupyterlab>=4.4.10",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2"
]

[project.optional-dependencies]
dev = [
    "black>=25.9.0",
    "ruff>=0.14.4",
    "pytest>=8.4.2",
    "ipykernel>=7.1.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "ml*", "rag*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"